+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Status condition to cluster
 
+### Fixed
+
+- Add finalizer to AWSCluster resource too
+
 ## [1.0.2] - 2022-09-23
 
 ### Fixed